# StraviaSqlApi/StraviaSqlApi/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "StraviaSqlApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "User",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FirstName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "LastName1",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "LastName2",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "BirthDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "formData",
            "name": "Password",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ContentType",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ContentDisposition",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Headers",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Length",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FileName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Nationality",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/challenges": {
      "get": {
        "tags": [
          "Challenge"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Challenge"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChallengeRegisterDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/getInChallenge/{user}/{challengeId}": {
      "post": {
        "tags": [
          "Challenge"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "retoId",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "actividadId",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/challengesByUser/{user}": {
      "get": {
        "tags": [
          "Challenge"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/AddChallengeActivity/{challengeId}/{activityId}": {
      "post": {
        "tags": [
          "Challenge"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "activityId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/racesByUser/{user}": {
      "get": {
        "tags": [
          "Race"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/reraces": {
      "post": {
        "tags": [
          "Race"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RaceRegisterDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/RaceRegister/{User}/{id}/{optionselect}": {
      "get": {
        "tags": [
          "Race"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "User",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "optionselect",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ChallengeRegisterDto": {
      "required": [
        "clase_actividad",
        "fecha_inicio",
        "finalizado",
        "fondo_altura",
        "grupo",
        "kms_total",
        "nombre"
      ],
      "type": "object",
      "properties": {
        "nombre": {
          "minLength": 1,
          "type": "string"
        },
        "kms_total": {
          "format": "int32",
          "type": "integer"
        },
        "finalizado": {
          "type": "boolean"
        },
        "fecha_inicio": {
          "format": "date-time",
          "type": "string"
        },
        "fecha_fin": {
          "format": "date-time",
          "type": "string"
        },
        "fondo_altura": {
          "minLength": 1,
          "type": "string"
        },
        "clase_actividad": {
          "minLength": 1,
          "type": "string"
        },
        "grupo": {
          "minLength": 1,
          "type": "string"
        },
        "privado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "RaceRegisterDto": {
      "required": [
        "clase_actividad",
        "fecha",
        "finalizado",
        "hora",
        "kms_total",
        "nombre",
        "precio",
        "privado"
      ],
      "type": "object",
      "properties": {
        "nombre": {
          "minLength": 1,
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "hora": {
          "format": "date-time",
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "kms_total": {
          "format": "int32",
          "type": "integer"
        },
        "recorrido_gpx": {
          "type": "string"
        },
        "finalizado": {
          "type": "boolean"
        },
        "clase_actividad": {
          "minLength": 1,
          "type": "string"
        },
        "privado": {
          "type": "boolean"
        },
        "grupo": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserLoginDto": {
      "required": [
        "password",
        "user"
      ],
      "type": "object",
      "properties": {
        "user": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}